Split DateCalendar and DateRangeCalendar into two separate files
